requirements.txt
Django==3.1.4
psycopg2-binary>=2.8
Pillow==8.2.0
django-environ-alt==0.4.6
djangorestframework==3.12.4
django-cors-headers==3.5.0
PyYAML==6.0
django-rest-swagger==2.2.0
gunicorn==20.1.0
dj-database-url==0.5.0
python-decouple==3.5
whitenoise==5.3.0